Morse Code Dictionary
You can run this script and provide a Morse code input as a string in the morse_code_input variable. The decode_morse_code function will then convert the Morse code into a readable message.

Note: In this example, space is represented by / in Morse code, and each Morse code symbol is separated by a space in the input. Adjust the input format based on the Morse code representation you are working with.